+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Add per-module option to add CUDA-C line information for profiling, use `wp.set_module_options({"lineinfo": True})`.
 - Add `example_tile_walker.py`, which reworks the existing `walker.py` to use Warp's tile API for matrix multiplication.
 - Add operator overloads for `wp.struct` objects by defining `wp.func` functions ([GH-392](https://github.com/NVIDIA/warp/issues/392)).
+- Add `example_tile_nbody.py`, an N-Body gravitational simulation example using Warp tile primitives.
 
 ### Changed